desktop_app/requirements.txt
PyQt5>=5.15.0
sounddevice>=0.4.5
soundfile>=0.10.3
numpy>=1.19.0
openai-whisper>=20231117
python-dotenv>=0.19.0
openpyxl>=3.0.9
pydantic>=1.8.0
groq>=0.0.1
